Lspinheiro/chore/migrate azure executor autogen ext (#3652)
* migrate code executor and tests

* update extras

* update dependencies and examples

* fix imports

* fix uv lock

* add code_executor to toctree

---------
