+++ python/packages/autogen-core/docs/src/reference/index.md
@@ -40,6 +40,7 @@ python/autogen_core/autogen_core
 
 python/autogen_ext/autogen_ext
 python/autogen_ext/autogen_ext.tools
+python/autogen_ext/autogen_ext.code_executor
 ```
 
 ::::{grid} 1 2 2 3
