+++ python/packages/autogen-core/src/autogen_core/components/code_executor/__init__.py
@@ -1,23 +1,37 @@
 from ._base import CodeBlock, CodeExecutor, CodeResult
-from ._func_with_reqs import Alias, FunctionWithRequirements, Import, ImportFromModule, with_requirements
-from ._impl.azure_container_code_executor import AzureContainerCodeExecutor
+from ._func_with_reqs import (
+    Alias,
+    FunctionWithRequirements,
+    FunctionWithRequirementsStr,
+    Import,
+    ImportFromModule,
+    build_python_functions_file,
+    to_stub,
+    with_requirements,
+)
 from ._impl.command_line_code_result import CommandLineCodeResult
 from ._impl.docker_command_line_code_executor import DockerCommandLineCodeExecutor
 from ._impl.local_commandline_code_executor import LocalCommandLineCodeExecutor
+from ._impl.utils import get_required_packages, lang_to_cmd
 from ._utils import extract_markdown_code_blocks
 
 __all__ = [
-    "AzureContainerCodeExecutor",
     "LocalCommandLineCodeExecutor",
     "CommandLineCodeResult",
     "CodeBlock",
-    "CodeResult",
     "CodeExecutor",
+    "CodeResult",
     "Alias",
     "ImportFromModule",
     "Import",
     "FunctionWithRequirements",
+    "FunctionWithRequirementsStr",
     "with_requirements",
+    "to_stub",
     "extract_markdown_code_blocks",
+    "get_required_packages",
+    "build_python_functions_file",
     "DockerCommandLineCodeExecutor",
+    "get_required_packages",
+    "lang_to_cmd",
 ]
