+++ python/packages/autogen-core/tests/execution/test_user_defined_functions.py
@@ -8,13 +8,11 @@ import polars
 import pytest
 from autogen_core.base import CancellationToken
 from autogen_core.components.code_executor import (
-    AzureContainerCodeExecutor,
     CodeBlock,
     FunctionWithRequirements,
     LocalCommandLineCodeExecutor,
     with_requirements,
 )
-from azure.identity import DefaultAzureCredential
 
 ENVIRON_KEY_AZURE_POOL_ENDPOINT = "AZURE_POOL_ENDPOINT"
 
@@ -78,32 +76,35 @@ print(data['name'][0])"""
         assert result.exit_code == 0
 
 
-@pytest.mark.skipif(
-    not POOL_ENDPOINT,
-    reason="do not run if pool endpoint is not defined",
-)
-@pytest.mark.asyncio
-async def test_azure_can_load_function_with_reqs() -> None:
-    assert POOL_ENDPOINT is not None
-    cancellation_token = CancellationToken()
-    azure_executor = AzureContainerCodeExecutor(
-        pool_management_endpoint=POOL_ENDPOINT, credential=DefaultAzureCredential(), functions=[load_data]
-    )
-    # AzureContainerCodeExecutor doesn't use the functions module import
-    code = """import polars
+def test_local_formatted_prompt() -> None:
+    assert_str = '''def add_two_numbers(a: int, b: int) -> int:
+    """Add two numbers together."""
+'''
+    with tempfile.TemporaryDirectory() as temp_dir:
+        executor = LocalCommandLineCodeExecutor(work_dir=temp_dir, functions=[add_two_numbers])
 
-# Get first row's name
-data = load_data()
-print(data['name'][0])"""
+        result = executor.format_functions_for_prompt()
+        assert assert_str in result
 
-    azure_result = await azure_executor.execute_code_blocks(
-        code_blocks=[
-            CodeBlock(language="python", code=code),
-        ],
-        cancellation_token=cancellation_token,
+
+def test_local_formatted_prompt_str_func() -> None:
+    func = FunctionWithRequirements.from_str(
+        '''
+def add_two_numbers(a: int, b: int) -> int:
+    """Add two numbers together."""
+    return a + b
+'''
     )
-    assert azure_result.output == "John\n"
-    assert azure_result.exit_code == 0
+
+    assert_str = '''def add_two_numbers(a: int, b: int) -> int:
+    """Add two numbers together."""
+'''
+
+    with tempfile.TemporaryDirectory() as temp_dir:
+        executor = LocalCommandLineCodeExecutor(work_dir=temp_dir, functions=[func])
+
+        result = executor.format_functions_for_prompt()
+        assert assert_str in result
 
 
 @pytest.mark.asyncio
@@ -124,31 +125,6 @@ print(add_two_numbers(1, 2))"""
         assert result.exit_code == 0
 
 
-@pytest.mark.skipif(
-    not POOL_ENDPOINT,
-    reason="do not run if pool endpoint is not defined",
-)
-@pytest.mark.asyncio
-async def test_azure_can_load_function() -> None:
-    assert POOL_ENDPOINT is not None
-
-    cancellation_token = CancellationToken()
-    azure_executor = AzureContainerCodeExecutor(
-        pool_management_endpoint=POOL_ENDPOINT, credential=DefaultAzureCredential(), functions=[add_two_numbers]
-    )
-    # AzureContainerCodeExecutor doesn't use the functions module import
-    code = """print(add_two_numbers(1, 2))"""
-
-    azure_result = await azure_executor.execute_code_blocks(
-        code_blocks=[
-            CodeBlock(language="python", code=code),
-        ],
-        cancellation_token=cancellation_token,
-    )
-    assert azure_result.output == "3\n"
-    assert azure_result.exit_code == 0
-
-
 @pytest.mark.asyncio
 async def test_fails_for_function_incorrect_import() -> None:
     with tempfile.TemporaryDirectory() as temp_dir:
@@ -166,30 +142,6 @@ function_incorrect_import()"""
             )
 
 
-@pytest.mark.skipif(
-    not POOL_ENDPOINT,
-    reason="do not run if pool endpoint is not defined",
-)
-@pytest.mark.asyncio
-async def test_azure_fails_for_function_incorrect_import() -> None:
-    assert POOL_ENDPOINT is not None
-    cancellation_token = CancellationToken()
-    azure_executor = AzureContainerCodeExecutor(
-        pool_management_endpoint=POOL_ENDPOINT,
-        credential=DefaultAzureCredential(),
-        functions=[function_incorrect_import],
-    )
-    code = """function_incorrect_import()"""
-
-    with pytest.raises(ValueError):
-        await azure_executor.execute_code_blocks(
-            code_blocks=[
-                CodeBlock(language="python", code=code),
-            ],
-            cancellation_token=cancellation_token,
-        )
-
-
 @pytest.mark.asyncio
 async def test_fails_for_function_incorrect_dep() -> None:
     with tempfile.TemporaryDirectory() as temp_dir:
@@ -205,73 +157,6 @@ function_incorrect_dep()"""
                 ],
                 cancellation_token=cancellation_token,
             )
-
-
-@pytest.mark.skipif(
-    not POOL_ENDPOINT,
-    reason="do not run if pool endpoint is not defined",
-)
-@pytest.mark.asyncio
-async def test_azure_fails_for_function_incorrect_dep() -> None:
-    assert POOL_ENDPOINT is not None
-    cancellation_token = CancellationToken()
-    azure_executor = AzureContainerCodeExecutor(
-        pool_management_endpoint=POOL_ENDPOINT, credential=DefaultAzureCredential(), functions=[function_incorrect_dep]
-    )
-    code = """function_incorrect_dep()"""
-
-    with pytest.raises(ValueError):
-        await azure_executor.execute_code_blocks(
-            code_blocks=[
-                CodeBlock(language="python", code=code),
-            ],
-            cancellation_token=cancellation_token,
-        )
-
-
-def test_formatted_prompt() -> None:
-    assert_str = '''def add_two_numbers(a: int, b: int) -> int:
-    """Add two numbers together."""
-'''
-    with tempfile.TemporaryDirectory() as temp_dir:
-        executor = LocalCommandLineCodeExecutor(work_dir=temp_dir, functions=[add_two_numbers])
-
-        result = executor.format_functions_for_prompt()
-        assert assert_str in result
-
-    azure_executor = AzureContainerCodeExecutor(
-        pool_management_endpoint=DUMMY_POOL_ENDPOINT, credential=DefaultAzureCredential(), functions=[add_two_numbers]
-    )
-
-    azure_result = azure_executor.format_functions_for_prompt()
-    assert assert_str in azure_result
-
-
-def test_formatted_prompt_str_func() -> None:
-    func = FunctionWithRequirements.from_str(
-        '''
-def add_two_numbers(a: int, b: int) -> int:
-    """Add two numbers together."""
-    return a + b
-'''
-    )
-
-    assert_str = '''def add_two_numbers(a: int, b: int) -> int:
-    """Add two numbers together."""
-'''
-
-    with tempfile.TemporaryDirectory() as temp_dir:
-        executor = LocalCommandLineCodeExecutor(work_dir=temp_dir, functions=[func])
-
-        result = executor.format_functions_for_prompt()
-        assert assert_str in result
-
-    azure_executor = AzureContainerCodeExecutor(
-        pool_management_endpoint=DUMMY_POOL_ENDPOINT, credential=DefaultAzureCredential(), functions=[func]
-    )
-
-    azure_result = azure_executor.format_functions_for_prompt()
-    assert assert_str in azure_result
 
 
 @pytest.mark.asyncio
@@ -298,36 +183,6 @@ print(add_two_numbers(1, 2))"""
         )
         assert result.output == "3\n"
         assert result.exit_code == 0
-
-
-@pytest.mark.skipif(
-    not POOL_ENDPOINT,
-    reason="do not run if pool endpoint is not defined",
-)
-@pytest.mark.asyncio
-async def test_azure_can_load_str_function_with_reqs() -> None:
-    assert POOL_ENDPOINT is not None
-    cancellation_token = CancellationToken()
-    func = FunctionWithRequirements.from_str(
-        '''
-def add_two_numbers(a: int, b: int) -> int:
-    """Add two numbers together."""
-    return a + b
-'''
-    )
-    azure_executor = AzureContainerCodeExecutor(
-        pool_management_endpoint=POOL_ENDPOINT, credential=DefaultAzureCredential(), functions=[func]
-    )
-    code = """print(add_two_numbers(1, 2))"""
-
-    azure_result = await azure_executor.execute_code_blocks(
-        code_blocks=[
-            CodeBlock(language="python", code=code),
-        ],
-        cancellation_token=cancellation_token,
-    )
-    assert azure_result.output == "3\n"
-    assert azure_result.exit_code == 0
 
 
 def test_cant_load_broken_str_function_with_reqs() -> None:
@@ -365,35 +220,3 @@ print(add_two_numbers(object(), False))"""
         )
         assert "TypeError: unsupported operand type(s) for +:" in result.output
         assert result.exit_code == 1
-
-
-@pytest.mark.skipif(
-    not POOL_ENDPOINT,
-    reason="do not run if pool endpoint is not defined",
-)
-@pytest.mark.asyncio
-async def test_azure_cant_run_broken_str_function_with_reqs() -> None:
-    assert POOL_ENDPOINT is not None
-    cancellation_token = CancellationToken()
-    func = FunctionWithRequirements.from_str(
-        '''
-def add_two_numbers(a: int, b: int) -> int:
-    """Add two numbers together."""
-    return a + b
-'''
-    )
-
-    azure_executor = AzureContainerCodeExecutor(
-        pool_management_endpoint=POOL_ENDPOINT, credential=DefaultAzureCredential(), functions=[func]
-    )
-    code = """print(add_two_numbers(object(), False))"""
-
-    azure_result = await azure_executor.execute_code_blocks(
-        code_blocks=[
-            CodeBlock(language="python", code=code),
-        ],
-        cancellation_token=cancellation_token,
-    )
-    # result.output = result.output.encode().decode('unicode_escape')
-    assert "TypeError: unsupported operand type(s) for +:" in azure_result.output
-    assert azure_result.exit_code == 1
